Handle case where current working directory is the removal target
- Add tests for core.remove_tests

+++ src/imgtrf/core.py
@@ -155,13 +155,19 @@ def _move_file(src_file_path: Path, dest_path: Path) -> None:
     dest_path.parent.mkdir(parents=True, exist_ok=True)
     shutil.move(src_file_path, dest_path)
 
-def remove_dirs(root_dir: Path) -> None:
 
+def remove_dirs(root_dir: Path) -> None:
+    """Recursivly remove empty directories"""
     for path in Path(root_dir).iterdir():
         if path.is_dir():
             remove_dirs(path)
+
+    if Path.cwd() == root_dir.resolve():
+        # Cant remove current working directory
+        return
 
     children = [child for child in root_dir.iterdir()]
     # remove if empty
     if len(children) == 0:
         root_dir.rmdir()
+        log.info(f"Removed {root_dir}")
